Refator: PR 리뷰에 대한 수정(1)
